删除无用多余的IPagedModel接口。
