+++ src/ServerApi/Infrastructures/Adnc.Infra.EfCore.MySQL/GlobalUsings.cs
@@ -1,5 +1,5 @@
 ﻿global using Adnc.Infra.Core.Interfaces;
-global using Adnc.Infra.EfCore;
+global using Adnc.Infra.EfCore.Transaction;
 global using Adnc.Infra.EfCore.Internal;
 global using Adnc.Infra.EfCore.MySQL;
 global using Adnc.Infra.EfCore.Repositories;
@@ -10,8 +10,8 @@ global using Microsoft.EntityFrameworkCore;
 global using Microsoft.EntityFrameworkCore.ChangeTracking;
 global using System;
 global using System.Collections.Generic;
-global using System.Diagnostics.CodeAnalysis;
 global using System.Linq;
 global using System.Reflection;
 global using System.Threading;
 global using System.Threading.Tasks;
+global using System.Diagnostics.CodeAnalysis;
