--- src/ServerApi/Infrastructures/Adnc.Infra.EfCore.MySQL/Interceptors/DefaultDbCommandInterceptor.cs
@@ -1,18 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore.Diagnostics;
-using System.Data.Common;
-
-namespace Adnc.Infra.EfCore.Interceptors
-{
-    public class DefaultDbCommandInterceptor : DbCommandInterceptor
-    {
-        public override ValueTask<InterceptionResult<DbDataReader>> ReaderExecutingAsync(DbCommand command, CommandEventData eventData, InterceptionResult<DbDataReader> result, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return base.ReaderExecutingAsync(command, eventData, result, cancellationToken);
-        }
-
-        public override ValueTask<InterceptionResult<int>> NonQueryExecutingAsync(DbCommand command, CommandEventData eventData, InterceptionResult<int> result, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            return base.NonQueryExecutingAsync(command, eventData, result, cancellationToken);
-        }
-    }
-}
