+++ src/ServerApi/Infrastructures/Adnc.Infra.EfCore.MySQL/Internal/EFCoreUtil.cs
@@ -1,8 +1,7 @@
-﻿namespace Adnc.Infra.EfCore.Internal
+﻿namespace Adnc.Infra.EfCore.Internal;
+
+internal static class EFCoreUtil
 {
-    public static class EFCoreUtil
-    {
-        public static object[] GetEntityKeyValues<TEntity>(Func<TEntity, object>[] keyValueGetter, TEntity entity)
-            => keyValueGetter.Select(x => x.Invoke(entity)).ToArray();
-    }
+    internal static object[] GetEntityKeyValues<TEntity>(Func<TEntity, object>[] keyValueGetter, TEntity entity)
+        => keyValueGetter.Select(x => x.Invoke(entity)).ToArray();
 }
