+++ src/ServerApi/Infrastructures/Adnc.Infra.EfCore.MySQL/Internal/KeyEntryModel.cs
@@ -1,11 +1,10 @@
-﻿namespace Adnc.Infra.EfCore.Internal
+﻿namespace Adnc.Infra.EfCore.Internal;
+
+public class KeyEntryModel
 {
-    public class KeyEntryModel
-    {
-        public string PropertyName { get; set; }
+    public string PropertyName { get; set; }
 
-        public string ColumnName { get; set; }
+    public string ColumnName { get; set; }
 
-        public object Value { get; set; }
-    }
+    public object Value { get; set; }
 }
