+++ src/ServerApi/Infrastructures/Adnc.Infra.EfCore.MySQL/Repositories/AbstractEfBaseRepository.cs
@@ -123,7 +123,7 @@ namespace Adnc.Infra.EfCore.Repositories
         protected virtual async Task<int> UpdateInternalAsync(TEntity entity, CancellationToken cancellationToken = default)
             => await DbContext.SaveChangesAsync(cancellationToken);
 
-        public virtual async Task<IPagedModel<TEntity>> PagedAsync(int pageIndex, int pageSize, Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, object>> orderByExpression, bool ascending = false, bool writeDb = false, CancellationToken cancellationToken = default)
+        public virtual async Task<PagedModel<TEntity>> PagedAsync(int pageIndex, int pageSize, Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, object>> orderByExpression, bool ascending = false, bool writeDb = false, CancellationToken cancellationToken = default)
         {
             var dbSet = this.GetDbSet(writeDb, false);
