+++ src/ServerApi/Infrastructures/Adnc.Infra.EfCore.MySQL/SqlGenerator/MySql/AdncQuerySqlGenerator.cs
@@ -7,37 +7,6 @@ using System.Linq.Expressions;
 
 namespace Pomelo.EntityFrameworkCore.MySql.Query.ExpressionVisitors.Internal
 {
-    /// <summary>
-    /// adnc sql生成工厂类
-    /// </summary>
-
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Usage", "EF1001:Internal EF Core API usage.", Justification = "<挂起>")]
-    public class AdncMySqlQuerySqlGeneratorFactory : MySqlQuerySqlGeneratorFactory
-    {
-        private readonly QuerySqlGeneratorDependencies _dependencies;
-        private readonly MySqlSqlExpressionFactory _sqlExpressionFactory;
-        private readonly IMySqlOptions _options;
-
-        public AdncMySqlQuerySqlGeneratorFactory(
-            [NotNull] QuerySqlGeneratorDependencies dependencies,
-            ISqlExpressionFactory sqlExpressionFactory,
-            IMySqlOptions options) : base(dependencies, options)
-        {
-            _dependencies = dependencies;
-            _sqlExpressionFactory = (MySqlSqlExpressionFactory)sqlExpressionFactory;
-            _options = options;
-        }
-
-        /// <summary>
-        /// 重写QuerySqlGenerator
-        /// </summary>
-        /// <returns></returns>
-        public override QuerySqlGenerator Create()
-        {
-            var result = new AdncQuerySqlGenerator(_dependencies, _sqlExpressionFactory, _options);
-            return result;
-        }
-    }
 
     /// <summary>
     /// adnc sql 生成类
