+++ src/ServerApi/Infrastructures/Adnc.Infra.EfCore.MySQL/Transaction/UnitOfWorkStatus.cs
@@ -1,0 +1,6 @@
+﻿namespace Adnc.Infra.EfCore.Transaction;
+
+public class UnitOfWorkStatus
+{
+    public bool IsStartingUow { get; internal set; }
+}
