--- src/ServerApi/Infrastructures/Adnc.Infra.EfCore.MySQL/UnitOfWork/UnitOfWorkStatus.cs
@@ -1,7 +1,0 @@
-﻿namespace Adnc.Infra.EfCore
-{
-    public class UnitOfWorkStatus
-    {
-        public bool IsStartingUow { get; internal set; }
-    }
-}
