--- src/ServerApi/Infrastructures/Adnc.Infra.Repository/IRepositories/Models/IPagedModel.cs
@@ -1,32 +1,0 @@
-﻿namespace Adnc.Infra.IRepositories
-{
-    public interface IPagedModel<out T>
-    {
-        /// <summary>
-        /// Data
-        /// </summary>
-        IReadOnlyList<T> Data { get; }
-
-        int Count { get; }
-
-        /// <summary>
-        /// PageNumber
-        /// </summary>
-        int PageIndex { get; }
-
-        /// <summary>
-        /// PageSize
-        /// </summary>
-        int PageSize { get; }
-
-        /// <summary>
-        /// TotalDataCount
-        /// </summary>
-        long TotalCount { get; set; }
-
-        /// <summary>
-        /// PageCount
-        /// </summary>
-        int PageCount { get; }
-    }
-}
