+++ src/ServerApi/Services/Adnc.Cus/Adnc.Cus.Application/AdncCusProfile.cs
@@ -4,7 +4,7 @@ public class AdncCusProfile : Profile
 {
     public AdncCusProfile()
     {
-        CreateMap(typeof(IPagedModel<>), typeof(PageModelDto<>)).ForMember("XData", opt => opt.Ignore());
+        CreateMap(typeof(PagedModel<>), typeof(PageModelDto<>)).ForMember("XData", opt => opt.Ignore());
         CreateMap<CustomerRegisterDto, Customer>();
         CreateMap<Customer, CustomerDto>();
     }
