+++ src/ServerApi/Services/Adnc.Ord/Adnc.Ord.Application/AdncOrdProfile.cs
@@ -4,7 +4,7 @@ public class AdncOrdProfile : Profile
 {
     public AdncOrdProfile()
     {
-        CreateMap(typeof(IPagedModel<>), typeof(PageModelDto<>)).ForMember("XData", opt => opt.Ignore());
+        CreateMap(typeof(PagedModel<>), typeof(PageModelDto<>)).ForMember("XData", opt => opt.Ignore());
 
         CreateMap<Order, OrderDto>();
         CreateMap<OrderItem, OrderDto.OrderItemDto>();
