+++ src/ServerApi/Services/Adnc.Whse/Adnc.Whse.Application/AdncWhseProfile.cs
@@ -4,7 +4,7 @@ public class AdncWhseProfile : Profile
 {
     public AdncWhseProfile()
     {
-        CreateMap(typeof(IPagedModel<>), typeof(PageModelDto<>)).ForMember("XData", opt => opt.Ignore());
+        CreateMap(typeof(PagedModel<>), typeof(PageModelDto<>)).ForMember("XData", opt => opt.Ignore());
 
         CreateMap<Product, ProductDto>();
